backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
aiomysql==0.2.0
alembic==1.14.0
PyMySQL==1.1.1

# Security — Authentication
python-jose[cryptography]==3.3.0   # JWT with RS256
argon2-cffi==23.1.0               # Argon2id password hashing (PHC winner)
cryptography==43.0.3              # AES-256-GCM, RSA key ops

# Input validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Rate limiting
slowapi==0.1.9
limits==3.13.0

# Utilities
python-multipart==0.0.12          # Form data
python-dotenv==1.0.1
httpx==0.28.1                     # Async HTTP client (weather API calls)
structlog==24.4.0                 # Structured JSON logging

# Hybrid fraud engine
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
lightgbm==4.5.0
catboost==1.2.7
shap==0.46.0
networkx==3.4.2
joblib==1.4.2
Pillow==10.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
